Filter added to get all users
