+++ src/controllers/UserController.js
@@ -2,6 +2,7 @@
 import Logger from '../loaders/logger';
 import Redis from '../middleware';
 import Service from '../services/index';
+import { Filter } from '../helpers';
 
 // eslint-disable-next-line import/named
 import DB from '../models';
@@ -9,7 +10,10 @@ import DB from '../models';
 const UserController = {
   async getAllUsers(req, res, next) {
     try {
-      const users = await DB.User.find({});
+      // const users = await DB.User.find({});
+      const users = await Filter.filter(req, 'users').then(
+        (response) => response,
+      );
       if (!users) {
         Logger.error('User was not found');
         return res.status(404).send('Nothing found');
